requirements.txt
anthropic>=0.39.0
requests>=2.31.0
GitPython>=3.1.40
structlog>=24.1.0
pytest>=7.4.3
ruff>=0.1.9
mypy>=1.7.1
pyyaml>=6.0.1
httpx>=0.25.2
tenacity>=8.2.3
